+++ Solution/Backend/dotnet/webapi/Repositories/Interfaces/IScoreRepository.cs
@@ -8,7 +8,7 @@ namespace backend.Repositories.Interfaces
     {
         Task<AverageScore> GetAverageScoreByIdAsync(int studentId);
         Task<List<AverageScore>> GetAllStudentsAverageScoreAsync();
-        Task<AssessmentScores> GetAssessmentResultData(int studentId, int assessmentId);
+        Task<AssessmentScores> GetAssessmentResultData(long studentId, long assessmentId);
 
     }
 }
